RESCRIS TEMA

+++ src/personal/homeworks/homework5/Ford.java
@@ -3,7 +3,7 @@ package personal.homeworks.homework5;
 public abstract class Ford extends Car {
     protected String model;
 
-    public Ford(int tankSize, FuelType fuelType, int gears, float consumption, float availableFuel, int tireSize, String chassisNumber, int currentGear, String model) {
+    public Ford(int tankSize, FuelType fuelType, int gears, double consumption, float availableFuel, int tireSize, String chassisNumber, int currentGear, String model) {
         super(tankSize, fuelType, gears, consumption, availableFuel, tireSize, chassisNumber, currentGear);
         this.model = model;
     }
